Drop - needless requirement for open ports
A service needn't wait for the iptables firewall rules to be changed (to
permit ingress connections) by lokkit before starting.  This
cross-module reference is only making other refactoring difficult, so it
is being eliminated here.  I've considered doing this for quite some
time, but lacked the justification for the effort.