New - iptables module
This new module subsumes the lokkit module for opening/closing TCP/UDP
ports, but goes further by fulling managing the iptables service.  This
is especially import in Fedora starting with 19 where the lokkit tool
does not enable the service, although admittedly I did not try its
--enable option.  It may also be that the iptables service is now just
disabled by default upon installation.  In any event, the service should
be restarted whenever the state is changed and while lokkit has done
this in the past, perhaps we shouldn't be relying on that behavior here.
Also, there is the future to consider where closing ports may be
desirable -- that doesn't work here due to limitations with lokkit.  We
may well be better off just inserting the appropriate lines ourselves...
or migrating to firewalld.  For now this makes Fedora 19 behave
consistently with prior releases however.