Bug - Fedora 32 doesn't have pam_krb5
I'm not even sure to what extent this was ever necessary or useful.
I can gain a console shell on hector-f32 or narvi-c7 without this module
in the PAM configuration and in all cases, I still get a Kerberos
ticket.  This may happen because of sssd, but I didn't investigate
further.

Signed-off-by: John Florian <jflorian@doubledog.org>